Fixed bug with data being stored to state and removed console logs in multiple places
