+++ src/ColorCard/ColorCard.js
@@ -4,7 +4,7 @@ import './ColorCard.scss';
 const ColorCard = () => {
     return(
         <section>
-            <img/> //to lock picture
+            <img/>
             <h3>COLOR OF PICTURE</h3>
         </section>
     )
